✨ feat: 사용자 정보 조회, 사용자 정보 수정 테스트 코드 합치기
